src/requirements.txt
Django==1.11.8
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
Pillow==7.1.1
pytz==2019.3
